Bug fixes and Updates - Bugs fixes, lu now can run successfully - Updates, add a hello example to show the basic stat - Code Formatting

+++ src/mem/mm.c
@@ -60,15 +60,15 @@ extern int repcnt[Setnum]; /* record the last replacement index of every set */
 
 
 int homehost(address_t addr){
-    return page[((unsigned long)(addr)-system_setting.global_start_addr) / Pagesize].homepid;
+    return page[((unsigned long long)(addr)-system_setting.global_start_addr) / Pagesize].homepid;
 }
 
 unsigned int homepage(address_t addr){
-    return page[((unsigned long)(addr)-system_setting.global_start_addr) / Pagesize].homei;
+    return page[((unsigned long long)(addr)-system_setting.global_start_addr) / Pagesize].homei;
 }
 
 unsigned int cachepage(address_t addr){
-    return page[((unsigned long)(addr)-system_setting.global_start_addr) / Pagesize].cachei;
+    return page[((unsigned long long)(addr)-system_setting.global_start_addr) / Pagesize].cachei;
 }
